Move out liquidation to it's own file
